pyproject.toml
[project]
name = "nonebot_plugin_furryfusion"
version = "0.1.0"
description = ""
authors = [
    {name = "Ekac", email = "2396382559@qq.com"},
]
dependencies = ["nonebot2", "httpx", "nonebot-adapter-onebot"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false
